requirements.txt
streamlit
google-generativeai
langchain
langchain-core
langchain-community
langchain-chroma
langchain-huggingface
sentence-transformers
chromadb
pypdf
duckduckgo-search
python-dotenv